feat(viewer): base the tour demo kit on building-architecture.ifc
Drop the fully-synthetic demo-project kit in favor of the committed
buildingSMART sample building-architecture.ifc. Compare and clash still
need authored content the base lacks, so a small rev B is DERIVED from
the base bytes (GlobalIds preserved) instead: it carries the four
compare deltas (data-modified LoadBearing, +0.5m moved wall, deleted
kitchen furniture, an added duct) and the added duct is positioned to
interpenetrate an existing wall by ~150mm, which doubles as the clash
tour's single hard clash (IfcDuctSegment vs IfcWall). A targeted
building-architecture.ids validates against the base with mixed
pass/fail grounded in real content (spaces named all-pass; walls
external 3/4; proxy ObjectType 3/5).

tools/demo-kit/generate.mts (synthetic IfcCreator model) is replaced by
the smaller derive-variants.mts, which reuses the STEP-surgery helpers
(delete scrubs dangling refs; add allocates ids above the closure max)
and self-verifies parse/diff/clash/IDS before writing. Both committed
models scan clean for duplicate ids and dangling references.

demo-kit.ts loader paths + manifest rekeyed to the building-architecture
names; the lens ruleset file is gone (the lens tour uses the builtin By
IFC Class preset).
