+++ AGENTS.md
@@ -55,7 +55,7 @@ Project-specific gotchas and guardrails — the things that bite you *here* and 
 
 ## Tour anchors & demo kit (viewer walkthroughs)
 - Elements carrying `data-tour` are interactive-tour anchors. If you move, rename, or delete one, update `apps/viewer/src/lib/tours/anchors.ts` and the referencing steps under `apps/viewer/src/lib/tours/tours/` in the same PR. A broken anchor does not fail CI: it auto-skips at runtime and fires a `tour_step_broken` PostHog event, so rot surfaces on the dashboard, not in review.
-- The demo kit (`apps/viewer/public/samples/demo-project*`) is generated by `tools/demo-kit/generate.mts` and committed. Revision B is DERIVED from the committed base bytes (diff matches on GlobalId, so a regenerated base makes compare read as 100% added/deleted). Never hand-edit or regenerate one artifact in isolation; rerun the generator, which self-verifies the IDS/clash/diff invariants.
+- The tour demo kit reuses the committed `building-architecture.ifc` sample as its base; its variants (`building-architecture-rev-b.ifc`, `building-architecture.ids`, `demo-kit.json`) are derived from that base by `tools/demo-kit/derive-variants.mts`. Revision B preserves the base's GlobalIds (diff matches on GlobalId, so a regenerated base would make compare read as 100% added/deleted) and carries the injected clash. Never hand-edit or regenerate one artifact in isolation; rerun the derivation script, which self-verifies the IDS/clash/diff invariants.
 
 ## New source files
 - MPL-2.0 header on every new file — see [`./LICENSE_HEADER.md`](./LICENSE_HEADER.md).
