+++ apps/viewer/public/samples/building-architecture.ids
@@ -1,37 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ids xmlns="http://standards.buildingsmart.org/IDS" xmlns:xs="http://www.w3.org/2001/XMLSchema" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://standards.buildingsmart.org/IDS http://standards.buildingsmart.org/IDS/1.0/ids.xsd">
   <info>
-    <title>Demo Project IDS</title>
-    <description>Sample buildingSMART IDS specifications for the ifc-lite interactive tour demo kit.</description>
+    <title>Building Architecture IDS</title>
+    <description>Sample buildingSMART IDS specifications for the ifc-lite interactive tour demo kit, grounded in the real "ifc silly sample scene" content.</description>
     <author>ifc-lite</author>
   </info>
   <specifications>
-    <specification name="Walls declare IsExternal" ifcVersion="IFC4">
-      <applicability>
-        <entity>
-          <name><simpleValue>IFCWALL</simpleValue></name>
-        </entity>
-      </applicability>
-      <requirements>
-        <property>
-          <propertySet><simpleValue>Pset_WallCommon</simpleValue></propertySet>
-          <baseName><simpleValue>IsExternal</simpleValue></baseName>
-        </property>
-      </requirements>
-    </specification>
-    <specification name="Walls have a fire rating" ifcVersion="IFC4">
-      <applicability>
-        <entity>
-          <name><simpleValue>IFCWALL</simpleValue></name>
-        </entity>
-      </applicability>
-      <requirements>
-        <property>
-          <propertySet><simpleValue>Pset_WallCommon</simpleValue></propertySet>
-          <baseName><simpleValue>FireRating</simpleValue></baseName>
-        </property>
-      </requirements>
-    </specification>
     <specification name="Spaces are named" ifcVersion="IFC4">
       <applicability>
         <entity>
@@ -41,6 +15,32 @@
       <requirements>
         <attribute>
           <name><simpleValue>Name</simpleValue></name>
+        </attribute>
+      </requirements>
+    </specification>
+    <specification name="Walls are external" ifcVersion="IFC4">
+      <applicability>
+        <entity>
+          <name><simpleValue>IFCWALL</simpleValue></name>
+        </entity>
+      </applicability>
+      <requirements>
+        <property dataType="IFCBOOLEAN">
+          <propertySet><simpleValue>Pset_WallCommon</simpleValue></propertySet>
+          <baseName><simpleValue>IsExternal</simpleValue></baseName>
+          <value><simpleValue>true</simpleValue></value>
+        </property>
+      </requirements>
+    </specification>
+    <specification name="Building element proxies declare an ObjectType" ifcVersion="IFC4">
+      <applicability>
+        <entity>
+          <name><simpleValue>IFCBUILDINGELEMENTPROXY</simpleValue></name>
+        </entity>
+      </applicability>
+      <requirements>
+        <attribute>
+          <name><simpleValue>ObjectType</simpleValue></name>
         </attribute>
       </requirements>
     </specification>
