+++ apps/viewer/src/lib/tours/demo-kit.ts
@@ -3,21 +3,25 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 /**
- * Demo-kit loaders. The kit (a purpose-built sample project plus fixtures
- * for the feature tours) is generated by `tools/demo-kit/generate.mts` and
- * committed under `public/samples/`; tours must never regenerate it at
- * runtime. Loading rides the existing `ifc-lite:load-file` bus event whose
- * listener routes into the normal loadFile pipeline (so `store.source`
- * bytes are retained for IDS et al).
+ * Demo-kit loaders. The kit demos the tours on the committed sample
+ * `building-architecture.ifc` plus two small variants derived from its
+ * bytes by `tools/demo-kit/derive-variants.mts` (a GlobalId-preserving
+ * revision B for compare, which also carries an injected hard clash, and a
+ * targeted IDS spec); tours must never regenerate them at runtime. Loading
+ * rides the existing `ifc-lite:load-file` bus event whose listener routes
+ * into the normal loadFile pipeline (so `store.source` bytes are retained
+ * for IDS et al).
  */
 
 import { EVENT_LOAD_FILE } from './events';
 
+const BASE_NAME = 'building-architecture.ifc';
+const REV_B_NAME = 'building-architecture-rev-b.ifc';
+
 export const DEMO_KIT_PATHS = {
-  base: '/samples/demo-project.ifc',
-  revB: '/samples/demo-project-rev-b.ifc',
-  ids: '/samples/demo-project.ids',
-  lenses: '/samples/demo-project.lenses.json',
+  base: `/samples/${BASE_NAME}`,
+  revB: `/samples/${REV_B_NAME}`,
+  ids: '/samples/building-architecture.ids',
   manifest: '/samples/demo-kit.json',
 } as const;
 
@@ -35,7 +39,11 @@ async function fetchAsFile(path: string, name: string): Promise<File> {
  * !geometryStreamingActive`) - same contract as a user-driven open.
  */
 export async function loadDemoProject(): Promise<void> {
-  const file = await fetchAsFile(DEMO_KIT_PATHS.base, 'demo-project.ifc');
+  const file = await fetchAsFile(DEMO_KIT_PATHS.base, BASE_NAME);
   // detail IS the File - the MainToolbar listener reads e.detail directly.
   window.dispatchEvent(new CustomEvent(EVENT_LOAD_FILE, { detail: file }));
 }
+
+/** The exact file names the demo variants load as (compare/IDS gate on
+ *  these to distinguish the kit from a user's own model). */
+export const DEMO_MODEL_NAMES = { base: BASE_NAME, revB: REV_B_NAME } as const;
